refactor: split 13 modules under the module-health caps (PR #3479 CI)
Extract cohesive concerns into new sibling modules so every file the
review-fix branch grew over the 500-LOC / 10-public-function caps returns
under cap, with mirrored behavior tests for each new module:

- backends/github/client.py 704->463 (pr_reads.py); gitlab/client.py 504->418
  (discussions.py); core/backend_factory.py 532->370 (toml_backends.py)
- core/management/commands/_test_plan/post.py 609->493 (mr_post.py)
- core/models/task.py 541->465 (task_claim.py, claim/lease CAS helpers)
- loop/scanners/pr_sweep.py 536->478 (pr_sweep_ports.py);
  slack_broadcasts.py 536->416 (slack_broadcast_mr_classifier.py)
- agents/headless.py 519->414 (headless_watchdog.py); core/speak.py 516->452
  (speak_cleaning.py)
- core/cleanup/cleanup.py 507->485 (cleanup_result.py);
  hooks/_body_file_resolution.py 515->401 (_inline_body_resolution.py);
  hooks/_command_parser.py 503->390 (_curl_payload.py)
- utils/git_commit.py: fold unsynced_commits_strict into unsynced_commits(strict=)
- core/management/commands/_shared_code_host.py: CodeHostErrorPayload TypedDict

Move the relocated deferred-import peg (backend_factory->toml_backends),
raise the flat-core pin 75->77 for the two new root siblings, and allowlist
_test_plan.mr_post in the forge-comment-write chokepoint (it scrubs via
route_forge_write before the raw backend call).
