+++ src/teatree/loop/scanners/__init__.py
@@ -49,11 +49,8 @@ from teatree.loop.scanners.review_request_merge_react import ReviewRequestMergeR
 from teatree.loop.scanners.reviewer_prs import ReviewerPrsScanner
 from teatree.loop.scanners.scanning_news import ScanningNewsScanner
 from teatree.loop.scanners.self_update import SelfUpdateScanner
-from teatree.loop.scanners.slack_broadcasts import (
-    BackendChannelHistoryFetcher,
-    GlabGhMrStateClassifier,
-    SlackBroadcastsScanner,
-)
+from teatree.loop.scanners.slack_broadcast_mr_classifier import GlabGhMrStateClassifier
+from teatree.loop.scanners.slack_broadcasts import BackendChannelHistoryFetcher, SlackBroadcastsScanner
 from teatree.loop.scanners.slack_dm_inbound import SlackDmInboundScanner
 from teatree.loop.scanners.slack_mentions import SlackMentionsScanner
 from teatree.loop.scanners.slack_review_intent import SlackReviewIntentScanner
