+++ tests/teatree_core/management/commands/test_shared_code_host.py
@@ -4,11 +4,16 @@ Eight commands used to open with divergent "no code host" wordings; this module
 is the one canonical message + the ``{"error": ...}`` payload built from it.
 """
 
-from teatree.core.management.commands._shared_code_host import NO_CODE_HOST_MESSAGE, no_code_host_error
+from teatree.core.management.commands._shared_code_host import (
+    NO_CODE_HOST_MESSAGE,
+    CodeHostErrorPayload,
+    no_code_host_error,
+)
 
 
 def test_no_code_host_error_wraps_the_canonical_message() -> None:
-    assert no_code_host_error() == {"error": NO_CODE_HOST_MESSAGE}
+    payload: CodeHostErrorPayload = no_code_host_error()
+    assert payload == {"error": NO_CODE_HOST_MESSAGE}
 
 
 def test_canonical_message_names_the_actionable_fix() -> None:
