+++ tests/teatree_utils/test_unsynced_commits.py
@@ -1,17 +1,17 @@
-"""``unsynced_commits_strict`` RAISES on a git failure instead of degrading to [].
+"""``unsynced_commits(strict=True)`` RAISES on a git failure instead of [].
 
-The lenient :func:`unsynced_commits` swallows a git failure to ``""`` → an empty
-list, which a destructive caller reads as "provably synced — safe to wipe". The
-strict twin (#F4.3) surfaces the failure as ``CommandFailedError`` so the caller
-can fail closed on an inconclusive probe, while still returning ``[]`` for a
-genuinely fully-synced branch.
+The lenient default swallows a git failure to ``""`` → an empty list, which a
+destructive caller reads as "provably synced — safe to wipe". Strict mode
+(#F4.3) surfaces the failure as ``CommandFailedError`` so the caller can fail
+closed on an inconclusive probe, while still returning ``[]`` for a genuinely
+fully-synced branch.
 """
 
 from pathlib import Path
 
 import pytest
 
-from teatree.utils.git import unsynced_commits, unsynced_commits_strict
+from teatree.utils.git import unsynced_commits
 from teatree.utils.git_run import run_strict
 from teatree.utils.run import CommandFailedError
 
@@ -39,19 +39,19 @@ def _repo_with_feature(tmp_path: Path, *, commits_ahead: int) -> Path:
 
 def test_lists_commits_ahead_of_base(tmp_path: Path) -> None:
     repo = _repo_with_feature(tmp_path, commits_ahead=2)
-    result = unsynced_commits_strict(str(repo), "feature", "main")
+    result = unsynced_commits(str(repo), "feature", "main", strict=True)
     assert len(result) == 2
 
 
 def test_empty_when_fully_synced(tmp_path: Path) -> None:
     repo = _repo_with_feature(tmp_path, commits_ahead=0)
-    assert unsynced_commits_strict(str(repo), "feature", "main") == []
+    assert unsynced_commits(str(repo), "feature", "main", strict=True) == []
 
 
-def test_raises_when_target_is_unresolvable(tmp_path: Path) -> None:
-    # An unresolvable target makes the underlying ``git log`` fail. The strict
-    # form RAISES; the lenient form silently returns [] (the data-loss trap).
+def test_strict_raises_but_lenient_swallows_when_target_is_unresolvable(tmp_path: Path) -> None:
+    # An unresolvable target makes the underlying ``git log`` fail. Strict mode
+    # RAISES; the lenient default silently returns [] (the data-loss trap).
     repo = _repo_with_feature(tmp_path, commits_ahead=1)
     with pytest.raises(CommandFailedError):
-        unsynced_commits_strict(str(repo), "feature", "origin/does-not-exist")
+        unsynced_commits(str(repo), "feature", "origin/does-not-exist", strict=True)
     assert unsynced_commits(str(repo), "feature", "origin/does-not-exist") == []
